Implement new 'rvslots' on MWAPI rv queries
Newer MediaWiki wikis provide warnings if 'rvslots' is not specified
while requesting revision information, while older ones warn if it is.

This change introduces a new 'slots' option for each wiki. If specified,
its value should be 'main' (as only the "main" slot is supported at this
time). 'rvslots' is used if the slots have been specified, otherwise the
older wiki behavior without 'rvslots' is used.

An example usage:

    wikis:
      - api: "https://wiki.cbhacks.com/api.php"
        s3_bucket: "wikiwatch.cbhacks.com"
        s3_prefix: "wiki.cbhacks.com/"
        slots: main
        sources:
          - generator: allpages

Non-"main" slots may be supported in the future